pyproject.toml
[tool.poetry]
name = "async-test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.95.1"}
sqlalchemy = "^2.0.9"
alembic = "^1.10.3"
python-dotenv = "^1.0.0"
asyncpg = "^0.27.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
bcrypt = "^4.0.1"
pyjwt = "^2.6.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
httpx = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
    ".", ".",
]
asyncio_mode="auto"